+++ frontend/app/players/players-table.tsx
@@ -2,7 +2,7 @@
 
 import { Table, TableHeader, TableColumn, TableBody, TableRow, TableCell, Link } from "@nextui-org/react";
 
-import { components } from "@/lib/battle-stadium-api";
+import { components } from "@/lib/api";
 
 type User = components["schemas"]["User"];
