+++ frontend/app/tournaments/page.tsx
@@ -5,7 +5,7 @@ import { auth } from "@clerk/nextjs/server";
 import { Card, CardBody, CardHeader } from "@nextui-org/react";
 
 import TournamentsTable from "@/components/tournaments-table";
-import { BattleStadiumAPI } from "@/lib/battle-stadium-api";
+import { BattleStadiumAPI } from "@/lib/api";
 
 export const metadata: Metadata = {
   title: "Tournaments",
