+++ frontend/components/sidebar/useSideBarItems.tsx
@@ -1,6 +1,6 @@
 // import { useUser } from "@clerk/nextjs";
 
-import { components } from "@/lib/battle-stadium-api";
+import { components } from "@/lib/api";
 
 import { SidebarItem } from "./sidebar";
 import { sectionItems } from "./sidebar-items";
