+++ frontend/lib/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./openapi-v1";
+export * from "./battle-stadium-api";
