--- frontend/lib/battle-stadium-api/index.ts
@@ -1,2 +1,0 @@
-export * from "./openapi-v1.d";
-export * from "./BattleStadiumAPI";
