+++ frontend/package.json
@@ -15,7 +15,7 @@
     "tsc:watch": "tsc --noEmit --watch",
     "fix": "bun prettier:write && bun lint:fix",
     "generate:factories": " ts-node --project ts-node.config.json generateFactories.ts",
-    "generate:api": "openapi-typescript ../backend/openapi/v1/openapi.yaml  -o ./lib/battle-stadium-api/openapi-v1.d.ts && bun fix",
+    "generate:api": "openapi-typescript ../backend/openapi/v1/openapi.yaml  -o ./lib/api/openapi-v1.d.ts && bun fix",
     "check:unused": "depcheck",
     "prisma:generate": "prisma generate",
     "prisma:pull": "prisma db pull"
